requirements.txt
werkzeug==2.3.7
python-dotenv==0.19.0
chromadb==0.4.0
pypdf==3.9.0
tiktoken==0.5.1
unstructured==0.10.30
google-generativeai==0.4.0
langchain-google-genai==0.0.12
groq==0.5.0
boto3==1.34.18
langchain==0.1.15
langchain-community==0.0.30
langchain-vectorstores==0.0.5
flask==3.0.0
